Project_8/P8_3_application_Flask_112022/requirements.txt
numpy==1.21.4
scikit-image==0.19.3
scikit-learn==1.0.2
seaborn==0.11.2
segmentation-models==1.0.1
pandas==1.3.4
tensorflow==2.10.0
Pillow==9.2.0
Flask==2.2.2


